Added GetUndoIndex()
This allows software using this widget to track file changes without saving a copy of the entire file.
Perplexed uses this to show whether a file has changes that can be saved, while maintaining the ability to undo/redo between saves.

This is necessary since `SetText(...)` clears the undo buffer.